+++ client/src/definitions/datasets.d.ts
@@ -1,0 +1,5 @@
+export interface Dataset {
+  id: string;
+  title: string;
+  description: string;
+}
