+++ client/src/definitions/fetch.ts
@@ -1,0 +1,1 @@
+export type Fetch = (request: Request) => Promise<Response>;
