+++ client/src/lib/components/DatasetList/DatasetList.svelte
@@ -1,0 +1,13 @@
+<script lang="ts">
+  import type { Dataset } from "src/definitions/datasets";
+
+  export let datasets: Dataset[];
+</script>
+
+<ul>
+  {#each datasets as dataset (dataset.id)}
+    <li>
+      {dataset.title}
+    </li>
+  {/each}
+</ul>
